authの判定を関数化
