+++ app/protected/page.tsx
@@ -1,20 +1,11 @@
 import DeployButton from "@/components/DeployButton";
 import AuthButton from "@/components/AuthButton";
-import { createClient } from "@/utils/supabase/server";
 import FetchDataSteps from "@/components/tutorial/FetchDataSteps";
 import Header from "@/components/Header";
-import { redirect } from "next/navigation";
+import { isAuthenticated } from "../products/fetcher";
 
 export default async function ProtectedPage() {
-  const supabase = createClient();
-
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
-  if (!user) {
-    return redirect("/login");
-  }
+  await isAuthenticated();
 
   return (
     <div className="flex-1 w-full flex flex-col gap-20 items-center">
